Implementation of the collection concept using a Book Library System
BOOK LIBRARY

Problem Description

A book library where a student, teacher borrow books. When the copy of the book is no longer in the library. The library should return ‘book taken’. They can be multiple copies of the same book in the library.

Implementation 1:

The books are given by the Librarian on a first come first serve basis, however, when a teacher is requesting for the same book a student is asking for, the teacher comes first, When a junior student is asking for the same book a senior student is asking for, the senior student comes first.

Implementation 2:

The books are given by the Librarian on a first come first serve basis, whether you are a teacher, a junior or a senior student.

How will I complete this project?

1. Write automated test

2. Organise the folders for your module (application), to house both your code base and the tests.

3. Write tests to cover all the methods to be written, before development begins (TDD).

4. Make logical assumptions where necessary.

Steps to evaluate.

1. Classes

a. At Minimum

i. UML should be used.

ii. The right visibility modifiers should be used.

iii. The following concepts/constructs should be used as much as possible

● Encapsulation

● Inheritance

● Polymorphism

● Interfaces/Abstract classes

● Single Responsibility for classes.

● Abstraction

● Aggregation

● Composition

● Generics

● Collections

● Exception Handling

iv. Use/Implement a PriorityQueue class where applicable.

2. Test Coverage

a. At Minimum

i. The tests should cover the methods as well as the conditions/procedures that the methods employed.